Fixing and cleaning

+++ rootItem.cs
@@ -94,7 +94,7 @@ namespace MusicStory
         private object authorField;
         private string langField;
         private link linkItemField;
-        private byte rateField;
+        private decimal rateField;
         private byte max_rateField;
         private byte percent_rateField;
         private object dateField;
@@ -367,7 +367,7 @@ namespace MusicStory
         public object author { get => authorField; set => authorField = value; }
         public string lang { get => langField; set => langField = value; }
         public link LinkItemField { get => linkItemField; set => linkItemField = value; }
-        public byte rate { get => rateField; set => rateField = value; }
+        public decimal rate { get => rateField; set => rateField = value; }
         public byte max_rate { get => max_rateField; set => max_rateField = value; }
         public byte percent_rate { get => percent_rateField; set => percent_rateField = value; }
         public object date { get => dateField; set => dateField = value; }
